refactor viewcontroller fix onDestroy in coroutines version
