+++ gradle.properties
@@ -17,5 +17,5 @@ org.gradle.jvmargs=-Xmx2048m -Dfile.encoding=UTF-8
 android.useAndroidX=true
 # Kotlin code style for this project: "official" or "obsolete":
 kotlin.code.style=official
-libraryVersion=6.0.3
+libraryVersion=6.0.4
 libraryGroup=ru.fabit.udf
